embed: Extra Recipes for Encoding Predictors

Predictors can be converted to one or more numeric representations using a variety of methods. Effect encodings using simple generalized linear models arXiv:1611.09477 or nonlinear models arXiv:1604.06737 can be used. There are also functions for dimension reduction and other approaches.

See also

Author

Maintainer: Emil Hvitfeldt emil.hvitfeldt@posit.co (ORCID)

Authors:

Max Kuhn max@posit.co (ORCID)

Other contributors:

Posit Software, PBC [copyright holder, funder]